remove the benchmark, ready for merging
